requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
requests==2.22.0
urllib3==1.25.7
six==1.13.0
pycparser==2.19
cffi==1.13.2
cryptography==2.8
future==0.18.2
certifi==2019.11.28
python-telegram-bot==12.2.0
tornado==6.0.3
peewee==3.13.1
